+++ windows/flutter/generated_plugins.cmake
@@ -7,7 +7,9 @@ list(APPEND FLUTTER_PLUGIN_LIST
   dynamic_color
   file_selector_windows
   flutter_secure_storage_windows
+  just_audio_windows
   permission_handler_windows
+  sqlite3_flutter_libs
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST